start-polling!: empty message queue before sleeping
This change ensures that the polling loop will do its best not to lag
behind with queued messages.

In issue #6, I proposed adding a procedure `poll-for-messages` to do
this kind of thing as a wrapper for `poll-for-message`.  However, the
implementation is so trivial for a dbus user to do on their own, I no
longer think that there is any need for that procedure in dbus egg.  (It
takes less effort to implement the fix in the calling program than it
does to understand the issue and distinction of an additional polling
procedure.)

Therefore, we implement the behavior in `start-polling!`, as that
procedure is just meant for the simplest uses of dbus, and I can't think
of a case where one would not want the polling thread to catch up with
the queue.

Resolves #6.