register-path, vtable: removed
`dbus_connection_register_object_path` is part of a different api for
interacting with DBus than the one this egg is using.  VTables are not
needed for the way this egg gets messages from DBus, namely
`dbus_connection_pop_message`.  The bindings here are also incorrect and
produced a compiler warning:

    Warning: in toplevel procedure `dbus2#make-vtable': (dbus2.scm:842)
      in procedure call to `dbus2#vtable-message_function-set!',
      expected argument #2 of type `(or boolean pointer locative)' but
      was given an argument of type `(procedure fn (* * *) fixnum)'

The simplest thing is to remove this code until such time as we decide
to rewrite the egg to use another method (like DBusWatch) to handle
messages.

This problem is also described here: https://bugs.call-cc.org/ticket/850

Resolves issue #5.